requirements.txt
python-telegram-bot
opencv-python
pillow
